v19 - prikaz volej pridan - jeste nefunguje

+++ src/Adventura/PrikazJdi.cpp
@@ -22,7 +22,7 @@ void PrikazJdi::proved(const std::vector<std::string> & parametry)
 		if (odkaz_na_plan->jdi(parametry[0]) == true)
 		{
 			auto aktualni_prostor = odkaz_na_plan->get_aktualni_prostor()->get_jmeno();
-			int mezer = 36 - aktualni_prostor.size();
+			int mezer = 36 - (int)aktualni_prostor.size();
 			std::cout <<
 				std::string("+-----------------------------------------------------------------------------+\n") +
 				std::string("|                          Jsi v prostoru ") +
